Refactor to use concern
