+++ app/mailers/event/application_paused_mailer.rb
@@ -6,7 +6,7 @@
 #  https://github.com/hitobito/hitobito_sac_cas.
 
 class Event::ApplicationPausedMailer < ApplicationMailer
-  include Rails.application.routes.url_helpers
+  include EventMailer
 
   NOTICE = "event_application_paused_notice"
 
@@ -15,40 +15,5 @@ class Event::ApplicationPausedMailer < ApplicationMailer
     locales = course.language.split("_")
 
     compose(course.groups.first.course_admin_email, NOTICE, {}, locales)
-  end
-
-  private
-
-  def placeholder_event_name
-    @course.name
-  end
-
-  def placeholder_event_number
-    @course.number
-  end
-
-  def placeholder_event_link
-    link_to "#{placeholder_event_name} (#{placeholder_event_number})",
-      group_event_url(group_id: @course.group_ids.first, id: @course.id)
-  end
-
-  def placeholder_event_details
-    info = []
-    info << labeled(:dates) { @course.dates.map(&:to_s).join("<br>") }
-    info << labeled(:motto)
-    info << labeled(:cost)
-    info << labeled(:description) { @course.description.gsub("\n", "<br>") }
-    info << labeled(:location) { @course.location.gsub("\n", "<br>") }
-    info << labeled(:contact) { "#{@course.contact}<br>#{@course.contact.email}" }
-    info.compact.join("<br><br>")
-  end
-
-  def labeled(key)
-    value = @course.send(key).presence
-    if value
-      label = @course.class.human_attribute_name(key)
-      formatted = block_given? ? yield : value
-      "#{label}:<br>#{formatted}"
-    end
   end
 end
